backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-multipart==0.0.9

PyMuPDF==1.24.10
python-docx==1.1.2

boto3==1.35.29
botocore==1.35.29
pydantic==2.9.2
python-dotenv==1.0.1